🚚 resource: organize files by routing
