+++ src/pages/Restaurants/CategoriesContainer.jsx
@@ -5,9 +5,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCategory,
   loadRestaurants,
-} from './slice';
+} from '../../redux/slice';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 import Buttons from './Buttons';
