+++ src/pages/Restaurants/RegionsContainer.jsx
@@ -5,9 +5,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from './slice';
+} from '../../redux/slice';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 import Buttons from './Buttons';
