+++ src/pages/Restaurants/RestaurantsContainer.jsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react';
 
 import { useSelector } from 'react-redux';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function RestaurantsContainer({ onClickRestaurant }) {
   const restaurants = useSelector(get('restaurants'));
